+++ CHANGELOG.md
@@ -26,6 +26,7 @@ and this project adheres to [Semantic Versioning].
   - added `__sub__` method
   - added `__mul__` and `__rmul__` method
   - added `conjugate` method
+  - added `inverse` method
 
 ## [0.0.1] - 2024-06-20